+++ src/main/java/com/skycatdev/antiscan/mixin/ServerLoginNetworkHandlerMixin.java
@@ -7,6 +7,7 @@ import com.skycatdev.antiscan.Utils;
 import net.minecraft.network.ClientConnection;
 import net.minecraft.network.packet.c2s.login.LoginHelloC2SPacket;
 import net.minecraft.server.network.ServerLoginNetworkHandler;
+import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -21,11 +22,15 @@ public abstract class ServerLoginNetworkHandlerMixin {
 
     @WrapMethod(method = "onHello")
     private void antiScan$tarpitBaddies(LoginHelloC2SPacket packet, Operation<Void> original) {
+        @Nullable String hostString = null;
+        if (connection.getAddress() instanceof InetSocketAddress inetSocketAddress) {
+            hostString = inetSocketAddress.getHostString();
+        }
         boolean good = switch (AntiScan.CONFIG.getLoginMode()) {
             case MATCH_EITHER -> {
-                if (connection.getAddress() instanceof InetSocketAddress inetSocketAddress) {
+                if (hostString != null) {
                     yield connection.isLocal() ||
-                          !AntiScan.IP_CHECKER.isBlacklisted(inetSocketAddress.getHostString()) ||
+                          !AntiScan.IP_CHECKER.isBlacklisted(hostString) ||
                           !AntiScan.NAME_CHECKER.isBlacklisted(packet.name());
                 }
                 yield connection.isLocal() || !AntiScan.NAME_CHECKER.isBlacklisted(packet.name());
@@ -33,15 +38,15 @@ public abstract class ServerLoginNetworkHandlerMixin {
             case MATCH_NONE -> true;
             case MATCH_ALL -> false;
             case MATCH_BOTH -> {
-                if (connection.getAddress() instanceof InetSocketAddress inetSocketAddress) {
+                if (hostString != null) {
                     yield connection.isLocal() ||
-                          !AntiScan.IP_CHECKER.isBlacklisted(inetSocketAddress.getHostString()) && !AntiScan.NAME_CHECKER.isBlacklisted(packet.name());
+                          !AntiScan.IP_CHECKER.isBlacklisted(hostString) && !AntiScan.NAME_CHECKER.isBlacklisted(packet.name());
                 }
                 yield connection.isLocal() && !AntiScan.NAME_CHECKER.isBlacklisted(packet.name());
             }
             case MATCH_IP -> {
-                if (connection.getAddress() instanceof InetSocketAddress inetSocketAddress) {
-                    yield connection.isLocal() || !AntiScan.IP_CHECKER.isBlacklisted(inetSocketAddress.getHostString());
+                if (hostString != null) {
+                    yield connection.isLocal() || !AntiScan.IP_CHECKER.isBlacklisted(hostString);
                 }
                 yield connection.isLocal();
             }
@@ -53,16 +58,18 @@ public abstract class ServerLoginNetworkHandlerMixin {
         } else {
             switch (AntiScan.CONFIG.getLoginAction()) {
                 case NOTHING -> original.call(packet);
-                case TARPIT -> {
+                case TARPIT -> AntiScan.LOGGER.info("Tarpitting {}.", hostString == null ? "connection" : hostString);
+                case DISCONNECT -> {
+                    AntiScan.LOGGER.info("Disconnecting {}.", hostString == null ? "connection" : hostString);
+                    connection.disconnect(Utils.translatable("multiplayer.disconnect.generic"));
                 }
-                case DISCONNECT -> connection.disconnect(Utils.translatable("multiplayer.disconnect.generic"));
                 default -> {
                     AntiScan.LOGGER.error("Impossible case - login action not handled. Allowing connection. Please report this at https://github.com/skycatminepokie/antiscan/issues.");
                     original.call(packet);
                 }
             }
-            if (AntiScan.CONFIG.isLoginReport() && connection.getAddress() instanceof InetSocketAddress inetSocketAddress) {
-                AntiScan.IP_CHECKER.report(inetSocketAddress.getHostString(), "Unsolicited connection attempt. Reported by AntiScan for Fabric.", new int[]{14});
+            if (AntiScan.CONFIG.isLoginReport() && hostString != null) {
+                AntiScan.IP_CHECKER.report(hostString, "Bad connection attempt. Reported by AntiScan for Fabric.", new int[]{14});
             }
         }
     }